requirements.txt
openpyxl==3.1.1
webcolors==1.13
robotpy-wpimath==2023.4.2.1
numpy==1.24.2
multipledispatch==0.6.0
pillow==9.4.0
git+https://github.com/nadegelemperiere/python-ldraw.git@v1.1.0